+++ src/net/Client.h
@@ -20,8 +20,8 @@ namespace net {
         /*
          * Sends a base64 encoded string to the server and returns a base64 encoded string as response.
          */
-        std::string communicate(std::string const &host, unsigned int port, std::string const &message);
-        std::string communicate(unsigned int port, std::string const &message);
+        std::string communicate(std::string const &host, unsigned short port, std::string const &message);
+        std::string communicate(unsigned short port, std::string const &message);
         std::string communicate(std::string const &host, std::string const &message);
         std::string communicate(std::string const &message);
     };
